+++ animals-api/Models/Animal.cs
@@ -10,10 +10,10 @@ namespace AnimalAPI.Models
     public class Animal
     {
         [Key]
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         [Required]
-        [StringLength(60, MinimumLength =3)]
+        [StringLength(60, MinimumLength = 3)]
         public string Name { get; set; }
 
         [Required]
